PRSD-623: Splits integration tests into IntegrationTests and JourneyIntegrationTests
