+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/InviteLaAdminTests.kt
@@ -1,7 +1,6 @@
 package uk.gov.communities.prsdb.webapp.integration
 
 import com.microsoft.playwright.Page
-import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
 import org.junit.jupiter.api.Test
 import org.mockito.ArgumentMatchers.anyString
 import org.mockito.kotlin.whenever
@@ -13,7 +12,7 @@ import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.B
 import java.net.URI
 
 @Sql("/data-local.sql")
-class InviteLaAdminTests : IntegrationTest() {
+class InviteLaAdminTests : JourneyIntegrationTest() {
     @Test
     fun `inviting a new LA admin ends with a confirmation page`(page: Page) {
         whenever(absoluteUrlProvider.buildInvitationUri(anyString()))
@@ -32,22 +31,5 @@ class InviteLaAdminTests : IntegrationTest() {
         assertPageIs(page, InviteLaAdminPage::class)
 
         // TODO PRSD-672 - check the Return to Dashboard button
-    }
-
-    @Test
-    fun `inviting a new LA admin shows validation errors if the email is invalid or the email addresses don't match`(page: Page) {
-        val invitePage = navigator.goToInviteLaAdmin()
-        invitePage.fillInFormAndSubmit("ISLE OF", "ISLE OF MAN", "not-an-email", "different@example.com")
-        assertThat(invitePage.form.getErrorMessage("email")).containsText("Enter an email address in the correct format")
-        assertThat(invitePage.form.getErrorMessage("confirmEmail")).containsText("Both email addresses should match")
-    }
-
-    @Test
-    fun `inviting a new LA admin shows validation errors if any of the fields are empty`(page: Page) {
-        val invitePage = navigator.goToInviteLaAdmin()
-        invitePage.form.submit()
-        assertThat(invitePage.form.getErrorMessage("localAuthorityId")).containsText("Select a local authority to continue")
-        assertThat(invitePage.form.getErrorMessage("email")).containsText("You must enter an email address")
-        assertThat(invitePage.form.getErrorMessage("confirmEmail")).containsText("You must enter and confirm their email address")
     }
 }
