+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/InviteLaUsersTests.kt
@@ -1,7 +1,6 @@
 package uk.gov.communities.prsdb.webapp.integration
 
 import com.microsoft.playwright.Page
-import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
 import org.junit.jupiter.api.Test
 import org.mockito.ArgumentMatchers.anyString
 import org.mockito.kotlin.whenever
@@ -13,7 +12,7 @@ import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.B
 import java.net.URI
 
 @Sql("/data-local.sql")
-class InviteLaUsersTests : IntegrationTest() {
+class InviteLaUsersTests : JourneyIntegrationTest() {
     @Test
     fun `inviting a new LA user ends with a success page with a button linking to the dashboard`(page: Page) {
         whenever(absoluteUrlProvider.buildInvitationUri(anyString()))
@@ -27,12 +26,5 @@ class InviteLaUsersTests : IntegrationTest() {
         // Go to dashboard button
         successPage.returnToDashboardButton.clickAndWait()
         assertPageIs(page, LocalAuthorityDashboardPage::class)
-    }
-
-    @Test
-    fun `inviting a new LA user shows validation errors if the email addresses don't match`() {
-        val invitePage = navigator.goToInviteNewLaUser(1)
-        invitePage.submitMismatchedEmails("test@example.com", "different@example.com")
-        assertThat(invitePage.form.getErrorMessage()).containsText("Both email addresses should match")
     }
 }
