+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/LandlordDeregistrationSinglePageTests.kt
@@ -1,0 +1,52 @@
+package uk.gov.communities.prsdb.webapp.integration
+
+import com.microsoft.playwright.Page
+import com.microsoft.playwright.assertions.PlaywrightAssertions
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.springframework.test.context.jdbc.Sql
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordDetailsPage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
+
+class LandlordDeregistrationSinglePageTests : IntegrationTest() {
+    @Nested
+    @Sql("/data-mockuser-landlord-with-properties.sql")
+    inner class LandlordWithProperties {
+        @Test
+        fun `User is returned to the landlord details page if they submit No`(page: Page) {
+            val areYouSurePage = navigator.goToLandlordDeregistrationAreYouSurePage()
+            areYouSurePage.submitDoesNotWantToProceed()
+
+            BasePage.assertPageIs(page, LandlordDetailsPage::class)
+        }
+
+        @Test
+        fun `User is returned to the landlord details page if they click the back link`(page: Page) {
+            val areYouSurePage = navigator.goToLandlordDeregistrationAreYouSurePage()
+            areYouSurePage.backLink.clickAndWait()
+            BasePage.assertPageIs(page, LandlordDetailsPage::class)
+        }
+
+        @Test
+        fun `Submitting with no option selected returns an error`() {
+            val areYouSurePage = navigator.goToLandlordDeregistrationAreYouSurePage()
+            areYouSurePage.form.submit()
+            PlaywrightAssertions
+                .assertThat(areYouSurePage.form.getErrorMessage("wantsToProceed"))
+                .containsText("Select whether you want to delete your landlord record and properties")
+        }
+    }
+
+    @Nested
+    @Sql("/data-unverified-landlord.sql")
+    inner class LandlordWithoutProperties {
+        @Test
+        fun `Submitting with no option selected returns an error`() {
+            val areYouSurePage = navigator.goToLandlordDeregistrationAreYouSurePage()
+            areYouSurePage.form.submit()
+            PlaywrightAssertions
+                .assertThat(areYouSurePage.form.getErrorMessage("wantsToProceed"))
+                .containsText("Select whether you want to delete your account from the database")
+        }
+    }
+}
