+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/PropertyRegistrationJourneyTests.kt
@@ -4,7 +4,6 @@ import com.microsoft.playwright.Page
 import com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.mockito.ArgumentCaptor.captor
 import org.mockito.kotlin.any
@@ -20,20 +19,15 @@ import uk.gov.communities.prsdb.webapp.constants.enums.PropertyType
 import uk.gov.communities.prsdb.webapp.database.entity.PropertyOwnership
 import uk.gov.communities.prsdb.webapp.database.repository.PropertyOwnershipRepository
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.assertThat
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.ErrorPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.LandlordDashboardPage
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage.Companion.assertPageIs
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyComplianceJourneyPages.StartPagePropertyCompliance
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.AlreadyRegisteredFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.CheckAnswersPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.ConfirmationPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.DeclarationFormPagePropertyRegistration
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.HmoAdditionalLicenceFormPagePropertyRegistration
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.HmoMandatoryLicenceFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.LicensingTypeFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.LookupAddressFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.ManualAddressFormPagePropertyRegistration
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.NoAddressFoundFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.NumberOfHouseholdsFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.NumberOfPeopleFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.OccupancyFormPagePropertyRegistration
@@ -43,7 +37,6 @@ import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyReg
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.SelectLocalAuthorityFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.SelectiveLicenceFormPagePropertyRegistration
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.TaskListPagePropertyRegistration
-import uk.gov.communities.prsdb.webapp.models.dataModels.AddressDataModel
 import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.emailModels.PropertyRegistrationConfirmationEmail
 import uk.gov.communities.prsdb.webapp.services.EmailNotificationService
@@ -51,7 +44,7 @@ import java.net.URI
 import kotlin.test.assertTrue
 
 @Sql("/data-local.sql")
-class PropertyRegistrationJourneyTests : IntegrationTest() {
+class PropertyRegistrationJourneyTests : JourneyIntegrationTest() {
     private val absoluteLandlordUrl = "www.prsd.gov.uk/landlord"
 
     @MockitoSpyBean
@@ -315,403 +308,5 @@ class PropertyRegistrationJourneyTests : IntegrationTest() {
         // Go to dashboard
         confirmationPage.goToDashboardButton.clickAndWait()
         assertPageIs(page, LandlordDashboardPage::class)
-    }
-
-    @Nested
-    inner class TaskListStep {
-        @Test
-        fun `Completing preceding steps will show a task as not yet started and completed steps as complete`(page: Page) {
-            navigator.skipToPropertyRegistrationOccupancyPage()
-            val taskListPage = navigator.goToPropertyRegistrationTaskList()
-            assert(taskListPage.taskHasStatus("Add the property address", "Complete"))
-            assert(taskListPage.taskHasStatus("Select the type of property", "Complete"))
-            assert(taskListPage.taskHasStatus("Select the ownership type", "Complete"))
-            assert(taskListPage.taskHasStatus("Add any property licensing information", "Complete"))
-            assert(taskListPage.taskHasStatus("Add any tenancy and household information", "Not yet started"))
-        }
-
-        @Test
-        fun `Completing first step of a task will show a task as in progress and completed steps as complete`(page: Page) {
-            navigator.skipToPropertyRegistrationHmoAdditionalLicencePage()
-            val taskListPage = navigator.goToPropertyRegistrationTaskList()
-            assert(taskListPage.taskHasStatus("Add the property address", "Complete"))
-            assert(taskListPage.taskHasStatus("Select the type of property", "Complete"))
-            assert(taskListPage.taskHasStatus("Select the ownership type", "Complete"))
-            assert(taskListPage.taskHasStatus("Add any property licensing information", "In progress"))
-            assert(taskListPage.taskHasStatus("Add any tenancy and household information", "Cannot start yet"))
-        }
-    }
-
-    @Nested
-    inner class LookupAddressAndNoAddressFoundSteps {
-        @Test
-        fun `Submitting with empty data fields returns an error`(page: Page) {
-            val lookupAddressPage = navigator.goToPropertyRegistrationLookupAddressPage()
-            lookupAddressPage.form.submit()
-            assertThat(lookupAddressPage.form.getErrorMessage("postcode")).containsText("Enter a postcode")
-            assertThat(lookupAddressPage.form.getErrorMessage("houseNameOrNumber")).containsText("Enter a house name or number")
-        }
-
-        @Test
-        fun `If no addresses are returned, user can search again or enter address manually via the No Address Found step`(page: Page) {
-            // Lookup address finds no results
-            val houseNumber = "15"
-            val postcode = "AB1 2CD"
-            whenever(osPlacesClient.search(houseNumber, postcode)).thenReturn("{}")
-            val lookupAddressPage = navigator.goToPropertyRegistrationLookupAddressPage()
-            lookupAddressPage.submitPostcodeAndBuildingNameOrNumber(postcode, houseNumber)
-
-            // redirect to noAddressFoundPage
-            val noAddressFoundPage = assertPageIs(page, NoAddressFoundFormPagePropertyRegistration::class)
-            assertThat(noAddressFoundPage.heading).containsText(houseNumber)
-            assertThat(noAddressFoundPage.heading).containsText(postcode)
-
-            // Search Again
-            noAddressFoundPage.searchAgain.clickAndWait()
-            val lookupAddressPageAgain = assertPageIs(page, LookupAddressFormPagePropertyRegistration::class)
-            lookupAddressPageAgain.submitPostcodeAndBuildingNameOrNumber(postcode, houseNumber)
-
-            // Submit no address found page
-            val noAddressFoundPageAgain = assertPageIs(page, NoAddressFoundFormPagePropertyRegistration::class)
-            noAddressFoundPageAgain.form.submit()
-            assertPageIs(page, ManualAddressFormPagePropertyRegistration::class)
-        }
-    }
-
-    @Nested
-    inner class SelectAddressStep {
-        @Test
-        fun `Submitting with no option selected returns an error`(page: Page) {
-            val selectAddressPage = navigator.skipToPropertyRegistrationSelectAddressPage()
-            selectAddressPage.form.submit()
-            assertThat(selectAddressPage.form.getErrorMessage()).containsText("Select an address")
-        }
-
-        @Test
-        fun `Clicking Search Again navigates to the previous step`(page: Page) {
-            val selectAddressPage = navigator.skipToPropertyRegistrationSelectAddressPage()
-            selectAddressPage.searchAgain.clickAndWait()
-            assertPageIs(page, LookupAddressFormPagePropertyRegistration::class)
-        }
-
-        @Test
-        fun `Selecting an already-registered address navigates to the AlreadyRegistered step`(page: Page) {
-            val alreadyRegisteredAddress = AddressDataModel("1 Example Road", uprn = 1123456)
-            val selectAddressPage = navigator.skipToPropertyRegistrationSelectAddressPage(listOf(alreadyRegisteredAddress))
-            selectAddressPage.selectAddressAndSubmit(alreadyRegisteredAddress.singleLineAddress)
-            assertPageIs(page, AlreadyRegisteredFormPagePropertyRegistration::class)
-        }
-    }
-
-    @Nested
-    inner class ManualAddressEntryStep {
-        @Test
-        fun `Submitting empty data fields returns errors`(page: Page) {
-            val manualAddressPage = navigator.skipToPropertyRegistrationManualAddressPage()
-            manualAddressPage.submitAddress()
-            assertThat(manualAddressPage.form.getErrorMessage("addressLineOne"))
-                .containsText("Enter the first line of an address, typically the building and street")
-            assertThat(manualAddressPage.form.getErrorMessage("townOrCity")).containsText("Enter town or city")
-            assertThat(manualAddressPage.form.getErrorMessage("postcode")).containsText("Enter postcode")
-        }
-    }
-
-    @Nested
-    inner class SelectLocalAuthorityStep {
-        @Test
-        fun `Submitting without selecting an LA return an error`(page: Page) {
-            val selectLocalAuthorityPage = navigator.skipToPropertyRegistrationSelectLocalAuthorityPage()
-            selectLocalAuthorityPage.form.submit()
-            assertThat(selectLocalAuthorityPage.form.getErrorMessage("localAuthorityId"))
-                .containsText("Select a local authority to continue")
-        }
-    }
-
-    @Nested
-    inner class PropertyTypeStep {
-        @Test
-        fun `Submitting with no propertyType selected returns an error`(page: Page) {
-            val propertyTypePage = navigator.skipToPropertyRegistrationPropertyTypePage()
-            propertyTypePage.form.submit()
-            assertThat(propertyTypePage.form.getErrorMessage()).containsText("Select the type of property")
-        }
-
-        @Test
-        fun `Submitting with the Other propertyType selected but an empty customPropertyType field returns an error`(page: Page) {
-            val propertyTypePage = navigator.skipToPropertyRegistrationPropertyTypePage()
-            propertyTypePage.submitCustomPropertyType("")
-            assertThat(propertyTypePage.form.getErrorMessage()).containsText("Enter the property type")
-        }
-    }
-
-    @Nested
-    inner class OwnershipTypeStep {
-        @Test
-        fun `Submitting with no ownershipType selected returns an error`(page: Page) {
-            val ownershipTypePage = navigator.skipToPropertyRegistrationOwnershipTypePage()
-            ownershipTypePage.form.submit()
-            assertThat(ownershipTypePage.form.getErrorMessage()).containsText("Select the ownership type")
-        }
-    }
-
-    @Nested
-    inner class LicensingTypeStep {
-        @Test
-        fun `Submitting with no licensingType selected returns an error`(page: Page) {
-            val licensingTypePage = navigator.skipToPropertyRegistrationLicensingTypePage()
-            licensingTypePage.form.submit()
-            assertThat(licensingTypePage.form.getErrorMessage()).containsText("Select the type of licensing for the property")
-        }
-
-        @Test
-        fun `Submitting with an HMO mandatory licence redirects to the next step`(page: Page) {
-            val licensingTypePage = navigator.skipToPropertyRegistrationLicensingTypePage()
-            licensingTypePage.submitLicensingType(LicensingType.HMO_MANDATORY_LICENCE)
-            val licenseNumberPage = assertPageIs(page, HmoMandatoryLicenceFormPagePropertyRegistration::class)
-            assertThat(licenseNumberPage.form.sectionHeader).containsText("Section 1 of 2 \u2014 Register your property details")
-        }
-
-        @Test
-        fun `Submitting with an HMO additional licence redirects to the next step`(page: Page) {
-            val licensingTypePage = navigator.skipToPropertyRegistrationLicensingTypePage()
-            licensingTypePage.submitLicensingType(LicensingType.HMO_ADDITIONAL_LICENCE)
-            val licenseNumberPage = assertPageIs(page, HmoAdditionalLicenceFormPagePropertyRegistration::class)
-            assertThat(licenseNumberPage.form.sectionHeader).containsText("Section 1 of 2 \u2014 Register your property details")
-        }
-    }
-
-    @Nested
-    inner class SelectiveLicenceStep {
-        @Test
-        fun `Submitting with no licence number returns an error`(page: Page) {
-            val selectiveLicencePage = navigator.skipToPropertyRegistrationSelectiveLicencePage()
-            selectiveLicencePage.form.submit()
-            assertThat(selectiveLicencePage.form.getErrorMessage()).containsText("Enter the selective licence number")
-        }
-
-        @Test
-        fun `Submitting with a very long licence number returns an error`(page: Page) {
-            val selectiveLicencePage = navigator.skipToPropertyRegistrationSelectiveLicencePage()
-            val aVeryLongString =
-                "This string is very long, so long that it is not feasible that it is a real licence number " +
-                    "- therefore if it is submitted there will in fact be an error rather than a successful submission." +
-                    " It is actually quite difficult for a string to be long enough to trigger this error, because the" +
-                    " maximum length has been selected to be permissive of id numbers we do not expect while still having " +
-                    "a cap reachable with a little effort."
-            selectiveLicencePage.submitLicenseNumber(aVeryLongString)
-            assertThat(selectiveLicencePage.form.getErrorMessage()).containsText("The licensing number is too long")
-        }
-    }
-
-    @Nested
-    inner class HmoMandatoryLicenceStep {
-        @Test
-        fun `Submitting with a licence number redirects to the next step`(page: Page) {
-            val hmoMandatoryLicencePage = navigator.skipToPropertyRegistrationHmoMandatoryLicencePage()
-            hmoMandatoryLicencePage.submitLicenseNumber("licence number")
-            assertPageIs(page, OccupancyFormPagePropertyRegistration::class)
-        }
-
-        @Test
-        fun `Submitting with no licence number returns an error`(page: Page) {
-            val hmoMandatoryLicencePage = navigator.skipToPropertyRegistrationHmoMandatoryLicencePage()
-            hmoMandatoryLicencePage.form.submit()
-            assertThat(hmoMandatoryLicencePage.form.getErrorMessage()).containsText("Enter the HMO Mandatory licence number")
-        }
-
-        @Test
-        fun `Submitting with a very long licence number returns an error`(page: Page) {
-            val hmoMandatoryLicencePage = navigator.skipToPropertyRegistrationHmoMandatoryLicencePage()
-            val aVeryLongString =
-                "This string is very long, so long that it is not feasible that it is a real licence number " +
-                    "- therefore if it is submitted there will in fact be an error rather than a successful submission." +
-                    " It is actually quite difficult for a string to be long enough to trigger this error, because the" +
-                    " maximum length has been selected to be permissive of id numbers we do not expect while still having " +
-                    "a cap reachable with a little effort."
-            hmoMandatoryLicencePage.submitLicenseNumber(aVeryLongString)
-            assertThat(hmoMandatoryLicencePage.form.getErrorMessage()).containsText("The licensing number is too long")
-        }
-    }
-
-    @Nested
-    inner class HmoAdditionalLicenceStep {
-        @Test
-        fun `Submitting with a licence number redirects to the next step`(page: Page) {
-            val hmoAdditionalLicencePage = navigator.skipToPropertyRegistrationHmoAdditionalLicencePage()
-            hmoAdditionalLicencePage.submitLicenseNumber("licence number")
-            assertPageIs(page, OccupancyFormPagePropertyRegistration::class)
-        }
-
-        @Test
-        fun `Submitting with no licence number returns an error`(page: Page) {
-            val hmoAdditionalLicencePage = navigator.skipToPropertyRegistrationHmoAdditionalLicencePage()
-            hmoAdditionalLicencePage.form.submit()
-            assertThat(hmoAdditionalLicencePage.form.getErrorMessage()).containsText("Enter the HMO additional licence number")
-        }
-
-        @Test
-        fun `Submitting with a very long licence number returns an error`(page: Page) {
-            val hmoAdditionalLicencePage = navigator.skipToPropertyRegistrationHmoAdditionalLicencePage()
-            val aVeryLongString =
-                "This string is very long, so long that it is not feasible that it is a real licence number " +
-                    "- therefore if it is submitted there will in fact be an error rather than a successful submission." +
-                    " It is actually quite difficult for a string to be long enough to trigger this error, because the" +
-                    " maximum length has been selected to be permissive of id numbers we do not expect while still having " +
-                    "a cap reachable with a little effort."
-            hmoAdditionalLicencePage.submitLicenseNumber(aVeryLongString)
-            assertThat(hmoAdditionalLicencePage.form.getErrorMessage()).containsText("The licensing number is too long")
-        }
-    }
-
-    @Nested
-    inner class OccupancyStep {
-        @Test
-        fun `Submitting with no occupancy option selected returns an error`(page: Page) {
-            val occupancyPage = navigator.skipToPropertyRegistrationOccupancyPage()
-            occupancyPage.form.submit()
-            assertThat(occupancyPage.form.getErrorMessage()).containsText("Select whether the property is occupied")
-        }
-    }
-
-    @Nested
-    inner class NumberOfHouseholdsStep {
-        @Test
-        fun `Submitting with a blank numberOfHouseholds field returns an error`(page: Page) {
-            val householdsPage = navigator.skipToPropertyRegistrationHouseholdsPage()
-            householdsPage.form.submit()
-            assertThat(householdsPage.form.getErrorMessage()).containsText("Enter the number of households living in your property")
-        }
-
-        @Test
-        fun `Submitting with a non-numerical value in the numberOfHouseholds field returns an error`(page: Page) {
-            val householdsPage = navigator.skipToPropertyRegistrationHouseholdsPage()
-            householdsPage.submitNumberOfHouseholds("not-a-number")
-            assertThat(householdsPage.form.getErrorMessage())
-                .containsText("Number of households in your property must be a positive, whole number, like 3")
-        }
-
-        @Test
-        fun `Submitting with a non-integer number in the numberOfHouseholds field returns an error`(page: Page) {
-            val householdsPage = navigator.skipToPropertyRegistrationHouseholdsPage()
-            householdsPage.submitNumberOfHouseholds("2.3")
-            assertThat(householdsPage.form.getErrorMessage())
-                .containsText("Number of households in your property must be a positive, whole number, like 3")
-        }
-
-        @Test
-        fun `Submitting with a negative integer in the numberOfHouseholds field returns an error`(page: Page) {
-            val householdsPage = navigator.skipToPropertyRegistrationHouseholdsPage()
-            householdsPage.submitNumberOfHouseholds(-2)
-            assertThat(householdsPage.form.getErrorMessage())
-                .containsText("Number of households in your property must be a positive, whole number, like 3")
-        }
-
-        @Test
-        fun `Submitting with a zero integer in the numberOfHouseholds field returns an error`(page: Page) {
-            val householdsPage = navigator.skipToPropertyRegistrationHouseholdsPage()
-            householdsPage.submitNumberOfHouseholds(0)
-            assertThat(householdsPage.form.getErrorMessage())
-                .containsText("Number of households in your property must be a positive, whole number, like 3")
-        }
-    }
-
-    @Nested
-    inner class NumberOfPeopleStep {
-        @Test
-        fun `Submitting with a blank numberOfPeople field returns an error`(page: Page) {
-            val peoplePage = navigator.skipToPropertyRegistrationPeoplePage()
-            peoplePage.form.submit()
-            assertThat(peoplePage.form.getErrorMessage()).containsText("Enter the number of people living in your property")
-        }
-
-        @Test
-        fun `Submitting with a non-numerical value in the numberOfPeople field returns an error`(page: Page) {
-            val peoplePage = navigator.skipToPropertyRegistrationPeoplePage()
-            peoplePage.submitNumOfPeople("not-a-number")
-            assertThat(peoplePage.form.getErrorMessage())
-                .containsText("Number of people in your property must be a positive, whole number, like 3")
-        }
-
-        @Test
-        fun `Submitting with a non-integer number in the numberOfPeople field returns an error`(page: Page) {
-            val peoplePage = navigator.skipToPropertyRegistrationPeoplePage()
-            peoplePage.submitNumOfPeople("2.3")
-            assertThat(peoplePage.form.getErrorMessage())
-                .containsText("Number of people in your property must be a positive, whole number, like 3")
-        }
-
-        @Test
-        fun `Submitting with a negative integer in the numberOfPeople field returns an error`(page: Page) {
-            val peoplePage = navigator.skipToPropertyRegistrationPeoplePage()
-            peoplePage.submitNumOfPeople("-2")
-            assertThat(peoplePage.form.getErrorMessage())
-                .containsText("Number of people in your property must be a positive, whole number, like 3")
-        }
-
-        @Test
-        fun `Submitting with a zero integer in the numberOfPeople field returns an error`(page: Page) {
-            val peoplePage = navigator.skipToPropertyRegistrationPeoplePage()
-            peoplePage.submitNumOfPeople(0)
-            assertThat(peoplePage.form.getErrorMessage())
-                .containsText("Number of people in your property must be a positive, whole number, like 3")
-        }
-
-        @Test
-        fun `Submitting with an integer in the numberOfPeople field that is less than the numberOfHouseholds returns an error`(
-            page: Page,
-        ) {
-            val householdsPage = navigator.skipToPropertyRegistrationHouseholdsPage()
-            householdsPage.submitNumberOfHouseholds(3)
-            val peoplePage = assertPageIs(page, NumberOfPeopleFormPagePropertyRegistration::class)
-            peoplePage.submitNumOfPeople(2)
-            assertThat(peoplePage.form.getErrorMessage())
-                .containsText(
-                    "The number of people in the property must be the same as or higher than the number of households in the property",
-                )
-        }
-    }
-
-    @Nested
-    inner class Declaration {
-        @Test
-        fun `Submitting without checking the checkbox returns an error`(page: Page) {
-            val declarationPage = navigator.skipToPropertyRegistrationDeclarationPage()
-            declarationPage.form.submit()
-            assertThat(declarationPage.form.getErrorMessage()).containsText("You must agree to the declaration to continue")
-        }
-    }
-
-    @Nested
-    inner class Confirmation {
-        @Test
-        fun `Navigating here with an incomplete form returns a 400 error page`(page: Page) {
-            navigator.navigateToPropertyRegistrationConfirmationPage()
-            val errorPage = assertPageIs(page, ErrorPage::class)
-            assertThat(errorPage.heading).containsText("Sorry, there is a problem with the service")
-        }
-    }
-
-    @Nested
-    inner class PropertyRegistrationStepCheckAnswers {
-        @Test
-        fun `After changing an answer, submitting a full section returns the CYA page`(page: Page) {
-            var checkAnswersPage = navigator.skipToPropertyRegistrationCheckAnswersPage()
-
-            checkAnswersPage.form.summaryList.ownershipRow.actions.actionLink
-                .clickAndWait()
-            val ownershipPage = assertPageIs(page, OwnershipTypeFormPagePropertyRegistration::class)
-
-            ownershipPage.submitOwnershipType(OwnershipType.LEASEHOLD)
-            checkAnswersPage = assertPageIs(page, CheckAnswersPagePropertyRegistration::class)
-
-            checkAnswersPage.form.summaryList.licensingRow.actions.actionLink
-                .clickAndWait()
-            val licensingTypePage = assertPageIs(page, LicensingTypeFormPagePropertyRegistration::class)
-
-            licensingTypePage.submitLicensingType(LicensingType.HMO_ADDITIONAL_LICENCE)
-            val licenceNumberPage = assertPageIs(page, HmoAdditionalLicenceFormPagePropertyRegistration::class)
-            licenceNumberPage.submitLicenseNumber("licence number")
-            assertPageIs(page, CheckAnswersPagePropertyRegistration::class)
-        }
     }
 }
